objects pkg reorg: consolidate RBAC packages, sharedconfig
Moves all RBAC-related object packages into
an rbac/ package under objects/. Also moves
the sharedconfig package contents up a level
into the existing objects package for simplicity.
