+++ internal/provisioner/objects/daemonset/daemonset.go
@@ -23,7 +23,6 @@ import (
 	"github.com/projectcontour/contour/internal/provisioner/labels"
 	"github.com/projectcontour/contour/internal/provisioner/model"
 	objutil "github.com/projectcontour/contour/internal/provisioner/objects"
-	objcfg "github.com/projectcontour/contour/internal/provisioner/objects/sharedconfig"
 
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
@@ -263,7 +262,7 @@ func DesiredDaemonSet(contour *model.Contour, contourImage, envoyImage string) *
 				"bootstrap",
 				filepath.Join("/", envoyCfgVolMntDir, envoyCfgFileName),
 				fmt.Sprintf("--xds-address=%s", contour.ContourServiceName()),
-				fmt.Sprintf("--xds-port=%d", objcfg.XDSPort),
+				fmt.Sprintf("--xds-port=%d", objutil.XDSPort),
 				fmt.Sprintf("--xds-resource-version=%s", xdsResourceVersion),
 				fmt.Sprintf("--resources-dir=%s", filepath.Join("/", envoyCfgVolMntDir, "resources")),
 				fmt.Sprintf("--envoy-cafile=%s", filepath.Join("/", envoyCertsVolMntDir, "ca.crt")),
