+++ internal/provisioner/objects/deployment/deployment.go
@@ -23,7 +23,6 @@ import (
 	"github.com/projectcontour/contour/internal/provisioner/labels"
 	"github.com/projectcontour/contour/internal/provisioner/model"
 	objutil "github.com/projectcontour/contour/internal/provisioner/objects"
-	objcfg "github.com/projectcontour/contour/internal/provisioner/objects/sharedconfig"
 
 	appsv1 "k8s.io/api/apps/v1"
 	corev1 "k8s.io/api/core/v1"
@@ -105,7 +104,7 @@ func EnsureDeploymentDeleted(ctx context.Context, cli client.Client, contour *mo
 // DesiredDeployment returns the desired deployment for the provided contour using
 // image as Contour's container image.
 func DesiredDeployment(contour *model.Contour, image string) *appsv1.Deployment {
-	xdsPort := objcfg.XDSPort
+	xdsPort := objutil.XDSPort
 	args := []string{
 		"serve",
 		"--incluster",
@@ -121,9 +120,9 @@ func DesiredDeployment(contour *model.Contour, image string) *appsv1.Deployment 
 	// Pass the insecure/secure flags to Contour if using non-default ports.
 	for _, port := range contour.Spec.NetworkPublishing.Envoy.ContainerPorts {
 		switch {
-		case port.Name == "http" && port.PortNumber != objcfg.EnvoyInsecureContainerPort:
+		case port.Name == "http" && port.PortNumber != objutil.EnvoyInsecureContainerPort:
 			args = append(args, fmt.Sprintf("--envoy-service-http-port=%d", port.PortNumber))
-		case port.Name == "https" && port.PortNumber != objcfg.EnvoySecureContainerPort:
+		case port.Name == "https" && port.PortNumber != objutil.EnvoySecureContainerPort:
 			args = append(args, fmt.Sprintf("--envoy-service-https-port=%d", port.PortNumber))
 		}
 	}
