+++ internal/provisioner/objects/service/service_test.go
@@ -19,7 +19,7 @@ import (
 	"testing"
 
 	"github.com/projectcontour/contour/internal/provisioner/model"
-	objcfg "github.com/projectcontour/contour/internal/provisioner/objects/sharedconfig"
+	objutil "github.com/projectcontour/contour/internal/provisioner/objects"
 
 	corev1 "k8s.io/api/core/v1"
 	"k8s.io/apimachinery/pkg/util/intstr"
@@ -141,7 +141,7 @@ func TestDesiredContourService(t *testing.T) {
 	}
 	cntr := model.New(cfg)
 	svc := DesiredContourService(cntr)
-	xdsPort := objcfg.XDSPort
+	xdsPort := objutil.XDSPort
 	checkServiceHasPort(t, svc, xdsPort)
 	checkServiceHasTargetPort(t, svc, xdsPort)
 	checkServiceHasPortName(t, svc, "xds")
